move contract to main
